Default to displaying only Active Users with an option to view All users
