+++ Plugins/BTCPayServer.RockstarDev.Plugins.Payroll/Data/Enums/PayrollUserState.cs
@@ -6,4 +6,12 @@
         Active,
         Archived
     }
+
+    public enum PayrollUserActiveState
+    {
+        All,
+        Disabled,
+        Active,
+        Archived
+    }
 }
