+++ Plugins/BTCPayServer.RockstarDev.Plugins.Payroll/Data/ViewModels/PayrollUser/PayrollUserCreateViewModel.cs
@@ -1,0 +1,27 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace BTCPayServer.RockstarDev.Plugins.Payroll.Data
+{
+    public class PayrollUserCreateViewModel
+    {
+        public string Id { get; set; }
+        [MaxLength(50)]
+        [Required]
+
+        public string Name { get; set; }
+
+        [Required]
+        [EmailAddress]
+        public string Email { get; set; }
+        [Required]
+        [MinLength(6)]
+        [DataType(DataType.Password)]
+        public string Password { get; set; }
+
+        [Required]
+        [DataType(DataType.Password)]
+        [Display(Name = "Confirm Password")]
+        [Compare("Password", ErrorMessage = "Password fields don't match")]
+        public string ConfirmPassword { get; set; }
+    }
+}
