+++ Plugins/BTCPayServer.RockstarDev.Plugins.Payroll/Data/ViewModels/PayrollUser/PayrollUserResetPasswordViewModel.cs
@@ -1,0 +1,20 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace BTCPayServer.RockstarDev.Plugins.Payroll.Data
+{
+    public class PayrollUserResetPasswordViewModel
+    {
+        public string Id { get; set; }
+        [Required]
+        [MinLength(6)]
+        [DataType(DataType.Password)]
+        [Display(Name = "New Password")]
+        public string NewPassword { get; set; }
+
+        [Required]
+        [DataType(DataType.Password)]
+        [Display(Name = "Confirm New Password")]
+        [Compare("NewPassword", ErrorMessage = "Password fields don't match")]
+        public string ConfirmNewPassword { get; set; }
+    }
+}
