+++ Plugins/BTCPayServer.RockstarDev.Plugins.Payroll/Views/PayrollUser/Edit.cshtml
@@ -1,5 +1,9 @@
 ﻿@using BTCPayServer.Abstractions.Extensions
 @using BTCPayServer.Abstractions.Contracts
+@using BTCPayServer.Abstractions.Models
+@using BTCPayServer.Client
+@using BTCPayServer.RockstarDev.Plugins.Payroll.Data
+@using BTCPayServer.RockstarDev.Plugins.Payroll.Data.Models
 @using BTCPayServer.Services
 @using static BTCPayServer.RockstarDev.Plugins.Payroll.Controllers.PayrollUserController
 @inject IScopeProvider ScopeProvider
